refactor : #266 OX 모션인식 결과 ENUM 추가
