+++ src/main/java/com/ssafy/freezetag/domain/balanceresult/entity/SelectedOption.java
@@ -1,0 +1,10 @@
+package com.ssafy.freezetag.domain.balanceresult.entity;
+
+public enum SelectedOption {
+    FIRST(0), SECOND(1);
+
+    private int value;
+    SelectedOption(final int value) {
+        this.value = value;
+    }
+}
