+++ src/main/java/com/ssafy/freezetag/domain/oxresult/entity/OXMotion.java
@@ -1,10 +1,10 @@
 package com.ssafy.freezetag.domain.oxresult.entity;
 
-public enum SelectedOption {
-    FIRST(0), SECOND(1);
+public enum OXMotion {
+    O(0), X(1), N(2);
 
     private int value;
-    SelectedOption(final int value) {
+    OXMotion(final int value) {
         this.value = value;
     }
 }
